refactor(poll_descriptor): split debug_log_throttle submodule (#4404 inc 1)
#4404 asks for the decomposition of poll_descriptor/mod.rs and its
poll_binding_process_descriptor god-function. The god-function itself
fuses the session-hit / session-miss / flowless arms and is guarded by
load-bearing invariants (single-recycle, CoS guarantee-guard,
packet-range/verifier-sensitive logic); per the issue it needs
/triple-review and is left untouched here. This is a bounded first
increment of pure code-motion.

Move the cold-path debug-log throttle unit out of mod.rs into a new
sibling submodule poll_descriptor/debug_log_throttle.rs:

  - the two #4120 interval caps SESSION_MISS_DEBUG_LOG_CAP and
    POLICY_DENY_DEBUG_LOG_CAP (kept module-private, only referenced
    inside the moved code),
  - the two pure u64 -> bool predicates session_miss_debug_log_allowed
    and policy_deny_debug_log_allowed (now pub(super) so the mod.rs call
    sites in poll_binding_process_descriptor still reach them),
  - the debug_log_throttle_tests pinned-contract test module (the #4120
    RED-on-revert guard that proves the throttle takes no topology
    input).

mod.rs gains a `mod debug_log_throttle;` declaration and a `use` import.
This unit was chosen because it has ZERO cross-references back into
mod.rs — the predicates depend on no dataplane type, so it is the
cleanest available boundary — and it directly matches the issue's stated
fix of outlining the cold-path debug logging. Behavior, the public
visibility surface, and all three test cases are identical; the bodies
move verbatim.

Validation: full `cargo test` (CARGO_TARGET_DIR=/tmp/cargo-4404) passes
across every binary — main lib 3739 passed / 0 failed (the 3 moved tests
now run as afxdp::poll_descriptor::debug_log_throttle::debug_log_throttle_tests::*),
plus 60 + 8 + 22 + 1 in the other binaries, 0 failed. The new file is
rustfmt-clean at edition 2024; mod.rs changed lines match surrounding
style. Module doc (mod.rs header) and _Log.md updated.
